test: move test methods around
